refactor(frontend): split lib/slack-api into a slack/ module
Move the converted Slack client into lib/slack/ (types.ts, client.ts,
and an index.ts facade), matching the lib/chat and lib/llm layout so
type definitions live apart from the API calls. Update the four
consumers and the test to import from @/lib/slack. The disconnectSlack
test now exercises the real 204 No Content response.
