+++ frontend/lib/slack/index.ts
@@ -1,0 +1,2 @@
+export * from "@/lib/slack/types";
+export * from "@/lib/slack/client";
